export

+++ API/SystemConnect.cs
@@ -99,20 +99,26 @@ namespace RocketForms.API
                 rtn += "<admin>";
 
                 var zipMapPath = _dataObject.AppTheme.ExportZipFile(_dataObject.PortalId, _dataObject.ModuleSettings.ModuleRef);
-                var systemByte = File.ReadAllBytes(zipMapPath);
-                var systemBase64 = Convert.ToBase64String(systemByte, Base64FormattingOptions.None);
-                rtn += "<systembase64 filetype='zip'><![CDATA[";
-                rtn += systemBase64;
-                rtn += "]]></systembase64>";
-                File.Delete(zipMapPath);
+                if (File.Exists(zipMapPath))
+                {
+                    var systemByte = File.ReadAllBytes(zipMapPath);
+                    var systemBase64 = Convert.ToBase64String(systemByte, Base64FormattingOptions.None);
+                    rtn += "<systembase64 filetype='zip'><![CDATA[";
+                    rtn += systemBase64;
+                    rtn += "]]></systembase64>";
+                    File.Delete(zipMapPath);
+                }
 
                 zipMapPath = _dataObject.AppTheme.ExportPortalZipFile(_dataObject.PortalId, _dataObject.ModuleSettings.ModuleRef);
-                systemByte = File.ReadAllBytes(zipMapPath);
-                systemBase64 = Convert.ToBase64String(systemByte, Base64FormattingOptions.None);
-                rtn += "<portalbase64 filetype='zip'><![CDATA[";
-                rtn += systemBase64;
-                rtn += "]]></portalbase64>";
-                File.Delete(zipMapPath);
+                if (File.Exists(zipMapPath))
+                {
+                    var systemByte = File.ReadAllBytes(zipMapPath);
+                    var systemBase64 = Convert.ToBase64String(systemByte, Base64FormattingOptions.None);
+                    rtn += "<portalbase64 filetype='zip'><![CDATA[";
+                    rtn += systemBase64;
+                    rtn += "]]></portalbase64>";
+                    File.Delete(zipMapPath);
+                }
 
                 rtn += "</admin>";
